+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+# 0.1.5 / 2023-12-15
+- Renamed `hxparams` to `hxvals` to match HTMX terminology.
+- Added hxwp_die() function to be used on templates (`noswap/` included). This functions will die() the script, but sending a 200 status code so HTMX can process the response and along with a header HX-Error on it, with the message included, so it can be used on the client side.
+-
+
 # 0.1.4 / 2023-12-13
 - Renamed `void/` endpoint to `noswap/` to match HTMX terminology, better showing the purpose of this endpoint.
 - Better path sanitization for template files.